refactor(shared): point db/ fragments at leaf modules, not the barrel (#814)
add-on/provider-access/renter-documents imported operators/users from
'./schema' (the barrel) and booking-types imported enum types from it too,
forming a fragment->barrel->fragment import cycle. Repoint operators/users
to './auth' (the post-#725 leaf) and booking-types' enum types to '../enums',
making the db/ import graph acyclic. Type-only + lazy FK thunks, so
behavior-preserving; db:generate no-op, shared 544/544.

Part of #701.
